+++ recipes/default.rb
@@ -19,5 +19,5 @@
 #
 
 vlc_app 'default' do
-  version node['vlc']['version'] if node['vlc']['version']
+  version node['vlc']['version']
 end