+++ spec/support/cookbooks/vlc_app_test/metadata.rb
@@ -1,0 +1,22 @@
+# Encoding: UTF-8
+#
+# rubocop:disable SingleSpaceBeforeFirstArg
+name             'vlc_app_test'
+maintainer       'Jonathan Hartman'
+maintainer_email 'j@p4nt5.com'
+license          'apache2'
+description      'vlc_app resource test wrapper'
+long_description 'vlc_app resource test wrapper'
+version          '0.0.1'
+
+depends          'vlc'
+
+supports         'mac_os_x'
+supports         'windows'
+supports         'ubuntu'
+supports         'debian'
+supports         'redhat'
+supports         'centos'
+supports         'scientific'
+supports         'freebsd'
+# rubocop:enable SingleSpaceBeforeFirstArg