+++ spec/support/cookbooks/vlc_app_test/recipes/default.rb
@@ -1,0 +1,5 @@
+# Encoding: UTF-8
+
+vlc_app 'default' do
+  version node['vlc']['app']['version']
+end